refactor(analysis): split archetype rules owner
