+++ docs/architecture-consolidation-plan.md
@@ -61,6 +61,7 @@ fixtures:
 | --- | --- | ---: | --- |
 | Content complexity | `crates/tokmd-analysis/src/content/complexity.rs` and `crates/tokmd-analysis/src/content/complexity/` | `tests/unit.rs` 1451; production owner modules <=187 | Scoring, nesting, and function-span helpers now live under owner modules; remaining work is mostly test split and aggregation cleanup |
 | Analysis orchestration | `crates/tokmd-analysis/src/analysis/mod.rs` and `crates/tokmd-analysis/src/analysis/` | coordinator 152; setup 20; files 53; outputs 22; enricher owners <=160 | Derived/source setup, file collection, optional output storage, and inventory/content/git/semantic/code-quality/effort enrichers now live in owner modules while `analyze` remains the public coordinator. Keep `analysis_orchestration` proof scoped to both the historical `analysis.rs` path and the new `analysis/**` owner paths during the file-to-folder transition |
+| Analysis archetype inference | `crates/tokmd-analysis/src/archetype/mod.rs` and `crates/tokmd-analysis/src/archetype/rules.rs` | coordinator 74; rule owner/tests 403 | Keep export-row filtering and path normalization in `mod.rs`; keep ordered Rust/Next.js/container/IaC/Python/Node archetype rules and rule tests in `rules.rs`. Preserve `analysis_security_identity` proof coverage. |
 | Analysis preset grid | `crates/tokmd-analysis/src/grid/mod.rs`, `grid/presets.rs`, and `grid/disabled_feature.rs` | `mod.rs` 10; preset matrix owner 445; disabled-feature warning owner 66 | Keep `mod.rs` as the public barrel; keep preset identity, preset-to-enricher matrix, and plan lookup in `presets.rs`; keep disabled-feature warning text in `disabled_feature.rs`. Preserve the `analysis_orchestration` proof scope. |
 | Analysis API surface | `crates/tokmd-analysis/src/api_surface/mod.rs` and `crates/tokmd-analysis/src/api_surface/` | `mod.rs` 13; report owner 227; symbol dispatcher 56; language scanners <=68; symbol tests 385 | Keep `mod.rs` as a thin coordinator, report aggregation in `report.rs`, source scanning in language owner modules under `symbols`, and leave large integration tests under `api_surface/tests` |
 | Analysis effort implementation | `crates/tokmd-analysis/src/effort/size_basis.rs`, `crates/tokmd-analysis/src/effort/classify.rs`, and `crates/tokmd-analysis/src/effort/classify/` | size-basis aggregation 271; file classification coordinator 80; `.gitattributes` owner/tests 145; heuristic owner/tests 215 | Authored/generated/vendored line aggregation stays in `size_basis.rs`; `classify.rs` coordinates explicit rules and fallback classification; `.gitattributes` parsing and path-pattern matching live in `classify/gitattributes.rs`; generated sentinels and heuristic file tagging live in `classify/heuristics.rs`. Keep this under the `analysis_effort` proof scope. |
